Add unit tests for expression conversion and enhance error handling
- Implemented ExpressionStructureTests to verify logical operations and method calls in expression conversion.
- Added MiscellaneousTests to check parameter name preservation during conversion.
- Created NestedPropertyMappingTests for testing nested property mappings.
- Introduced ScenarioTests for real-world conversion scenarios with comprehensive member mappings.
- Enhanced ExpressionConverter to support custom error handling strategies for member mapping failures.
- Introduced MemberMappingErrorHandling enum to specify error handling behavior.
- Updated ExpressionConverterOptions to allow fluent configuration of error handling, member mapping, and custom mapping.
- Removed obsolete package license URL from project file.
