+++ CrossTypeExpressionConverter.Tests/Units/ScenarioTests.cs
@@ -1,0 +1,82 @@
+using System;
+using System.Linq.Expressions;
+using CrossTypeExpressionConverter.Tests.Helpers.Models;
+
+namespace CrossTypeExpressionConverter.Tests.Units;
+
+/// <summary>
+/// Contiene tests para escenarios de conversión del mundo real.
+/// </summary>
+[TestFixture]
+public class ScenarioTests
+{
+    private bool Evaluate<T>(Expression<Func<T, bool>> predicate, T item)
+    {
+        return predicate.Compile()(item);
+    }
+
+    [Test]
+    public void Convert_NewsToArticle_WithComprehensiveMemberMap_TitleMatch_ShouldEvaluateCorrectly()
+    {
+        // Arrange
+        var newsToArticleMap = new Dictionary<string, string>
+        {
+            { nameof(NewsDataModel.UserIdOwner), nameof(ArticleDataModel.OwnerId) },
+            { nameof(NewsDataModel.NewsTitle), nameof(ArticleDataModel.Title) },
+            { nameof(NewsDataModel.NewsContent), nameof(ArticleDataModel.Content) },
+            // ... otros mapeos ...
+        };
+        var options = new ExpressionConverterOptions().WithMemberMap(newsToArticleMap);
+        Expression<Func<NewsDataModel, bool>> sourcePredicate = news => news.NewsTitle == "Breaking News";
+
+        // Act
+        var convertedPredicate = ExpressionConverter.Convert<NewsDataModel, ArticleDataModel>(sourcePredicate, options);
+
+        // Assert
+        Assert.IsTrue(Evaluate(convertedPredicate, new ArticleDataModel { Title = "Breaking News" }));
+        Assert.IsFalse(Evaluate(convertedPredicate, new ArticleDataModel { Title = "Old News" }));
+    }
+
+    [Test]
+    public void Convert_NewsToArticle_WithComprehensiveMemberMap_DateComparison_ShouldEvaluateCorrectly()
+    {
+        // Arrange
+        var newsToArticleMap = new Dictionary<string, string> 
+        {
+            { nameof(NewsDataModel.PublicationDate), nameof(ArticleDataModel.PublicationDate) },
+             // ... otros mapeos ...
+        };
+        var options = new ExpressionConverterOptions().WithMemberMap(newsToArticleMap);
+        var testDate = new DateTime(2023, 1, 15);
+        Expression<Func<NewsDataModel, bool>> sourcePredicate = news => news.PublicationDate > testDate;
+
+        // Act
+        var convertedPredicate = ExpressionConverter.Convert<NewsDataModel, ArticleDataModel>(sourcePredicate, options);
+
+        // Assert
+        Assert.IsTrue(Evaluate(convertedPredicate, new ArticleDataModel { PublicationDate = testDate.AddDays(1) }));
+        Assert.IsFalse(Evaluate(convertedPredicate, new ArticleDataModel { PublicationDate = testDate.AddDays(-1) }));
+    }
+
+    [Test]
+    public void Convert_SubjectToDatamodel_NameAndMandatory_WithMemberMap_ShouldEvaluateCorrectly()
+    {
+        // Arrange
+        var subjectMap = new Dictionary<string, string>
+        {
+            { nameof(SubjectDatabaseModel.SubjectName), nameof(SubjectsDatamodel.NameOfSubject) },
+            { nameof(SubjectDatabaseModel.IsMandatory), nameof(SubjectsDatamodel.Mandatory) }
+        };
+        var options = new ExpressionConverterOptions().WithMemberMap(subjectMap);
+        Expression<Func<SubjectDatabaseModel, bool>> sourcePredicate = 
+            subject => subject.SubjectName.Contains("Math") && subject.IsMandatory;
+            
+        // Act
+        var convertedPredicate = ExpressionConverter.Convert<SubjectDatabaseModel, SubjectsDatamodel>(sourcePredicate, options);
+
+        // Assert
+        Assert.IsTrue(Evaluate(convertedPredicate, new SubjectsDatamodel { NameOfSubject = "Advanced Mathematics", Mandatory = true }));
+        Assert.IsFalse(Evaluate(convertedPredicate, new SubjectsDatamodel { NameOfSubject = "History", Mandatory = true }));
+        Assert.IsFalse(Evaluate(convertedPredicate, new SubjectsDatamodel { NameOfSubject = "Basic Mathematics", Mandatory = false }));
+    }
+}
